feat: 경로 조회하는 로직 구현 완료

+++ src/main/java/wooteco/subway/dto/response/PathResponse.java
@@ -6,10 +6,10 @@ import wooteco.subway.domain.Station;
 public class PathResponse {
 
     private final List<Station> stations;
-    private final Long distance;
+    private final int distance;
     private final int fare;
 
-    public PathResponse(List<Station> stations, Long distance, int fare) {
+    public PathResponse(List<Station> stations, int distance, int fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -19,7 +19,7 @@ public class PathResponse {
         return stations;
     }
 
-    public Long getDistance() {
+    public int getDistance() {
         return distance;
     }
